Added config files * json and yaml versions of default values from MOSGUITO * experiments.xlsx Name generation in Snakefile now is not called if name already exists * also removed config.json call

+++ workflow/scripts/annotation.py
@@ -13,42 +13,42 @@ import multiprocessing
 import numpy as np
 import os
 import pandas as pd
-from mosca_tools import run_command, fastq2fasta
+from .mosca_tools import run_command, fastq2fasta
 from progressbar import ProgressBar
 
 
 class Annotater:
-    
-    def __init__ (self, **kwargs):
+
+    def __init__(self, **kwargs):
         self.__dict__ = kwargs
-        
-    def get_arguments(self):    
+
+    def get_arguments(self):
         parser = argparse.ArgumentParser(description="MOSCA annotation")
-        parser.add_argument("-i", "--input", type = str, required = True,
+        parser.add_argument("-i", "--input", type=str, required=True,
                             help="""Can be filename of single-end reads,
                             comma separated-list of filenames of paired-end reads
                             or filename of contigs file (this one requires the 
                             "--assembled option".""")
-        parser.add_argument("-t", "--threads", type = str, 
-                            default = str(multiprocessing.cpu_count() - 2),
-                            help = "Number of threads to use. Default is number of CPUs available minus 2.")
-        parser.add_argument("-a", "--assembled", action = "store_true", default = False,
+        parser.add_argument("-t", "--threads", type=str,
+                            default=str(multiprocessing.cpu_count() - 2),
+                            help="Number of threads to use. Default is number of CPUs available minus 2.")
+        parser.add_argument("-a", "--assembled", action="store_true", default=False,
                             help="If input is assembled reads.")
-        parser.add_argument("-o", "--output", type = str, help = "Output directory"),
-        parser.add_argument("-em", "--error-model", type = str, default = 'illumina_5',
-                            help = "Error model for FastQ reads input", 
-                            choices = ['sanger_5', 'sanger_10', '454_10', '454_30',
-                                       'illumina_5', 'illumina_10'])
-        parser.add_argument("-db", "--database", type = str, help = "Database for annotation")
+        parser.add_argument("-o", "--output", type=str, help="Output directory"),
+        parser.add_argument("-em", "--error-model", type=str, default='illumina_5',
+                            help="Error model for FastQ reads input",
+                            choices=['sanger_5', 'sanger_10', '454_10', '454_30',
+                                     'illumina_5', 'illumina_10'])
+        parser.add_argument("-db", "--database", type=str, help="Database for annotation")
         parser.add_argument("-mts", "--max-target-seqs", type=str, help="Number of identifications for each protein")
-        parser.add_argument("--download-uniprot", action = "store_true", default = False,
+        parser.add_argument("--download-uniprot", action="store_true", default=False,
                             help="Download uniprot database if FASTA DB doesn't exist")
-    
+
         args = parser.parse_args()
-        
+
         args.output = args.output.rstrip('/')
         return args
-        
+
     '''
     Input:
         file: name of input file to perform gene calling on
@@ -61,56 +61,70 @@ class Annotater:
         be produced in the same folder with the same name as the FASTQ file
         (but with .fasta instead of .fastq)
     '''
-    def gene_calling(self, file, output, threads = '12', assembled = True, 
-                     error_model = 'illumina_10'):
+
+    def gene_calling(self, file, output, threads='12', assembled=True,
+                     error_model='illumina_10'):
         if not assembled:
-            fastq2fasta(file, file.replace('fastq', 'fasta'))                   # FragGeneScan only accepts as input a FASTA file
-        
+            fastq2fasta(file, file.replace('fastq', 'fasta'))  # FragGeneScan only accepts as input a FASTA file
+
         run_command('run_FragGeneScan.pl -thread={} -genome={}'.format(threads,
-            '{} -out={} -complete=1 -train=./complete'.format(file, output) if assembled
-            else '{} -out={} -complete=0 -train=./'.format(file.replace(
-                'fastq', 'fasta'), error_model)))
+                                                                       '{} -out={} -complete=1 -train=./complete'.format(
+                                                                           file, output) if assembled
+                                                                       else '{} -out={} -complete=0 -train=./'.format(
+                                                                           file.replace(
+                                                                               'fastq', 'fasta'), error_model)))
 
     def download_uniprot(self, out_dir):
-        for db in ['trembl', 'sprot']:
-            run_command('wget ftp://ftp.uniprot.org/pub/databases/uniprot/current_release/knowledgebase/complete/uniprot_{}.fasta.gz'.format(db))
-        run_command('zcat uniprot_trembl.fasta.gz uniprot_sprot.fasta.gz', file='{}/uniprot.fasta'.format(out_dir))
-        run_command('rm uniprot_trembl.fasta.gz uniprot_sprot.fasta.gz')
+        if not os.path.isfile('{}/uniprot.fasta'.format(out_dir)):
+            for db in ['trembl', 'sprot']:
+                run_command(
+                    'wget ftp://ftp.uniprot.org/pub/databases/uniprot/current_release/knowledgebase/complete/uniprot_{}.fasta.gz'.format(
+                        db))
+            run_command('zcat uniprot_trembl.fasta.gz uniprot_sprot.fasta.gz', file='{}/uniprot.fasta'.format(out_dir))
+            run_command('rm uniprot_trembl.fasta.gz uniprot_sprot.fasta.gz')
+        else:
+            print('UniProt database found at: {}/uniprot.fasta'.format(out_dir))
 
     def generate_diamond_database(self, fasta, dmnd):
         run_command('diamond makedb --in {} -d {}'.format(fasta, dmnd))
 
-    def run_diamond(self, query, aligned, unaligned, database, threads = '12', 
-                    max_target_seqs = '50'):
+    def run_diamond(self, query, aligned, unaligned, database, threads='12',
+                    max_target_seqs='50'):
         if database[-6:] == '.fasta' or database[-4:] == '.faa':
             print('FASTA database was inputed')
-            if not os.path.isfile(database.replace('fasta','dmnd')):
+            if not os.path.isfile(database.replace('fasta', 'dmnd')):
                 print('DMND database not found. Generating a new one')
-                self.generate_diamond_database(database, 
-                                               database.replace('fasta','dmnd'))
+                self.generate_diamond_database(database,
+                                               database.replace('fasta', 'dmnd'))
             else:
                 print('DMND database was found. Using it')
             database = database.split('.fa')[0]
         elif database[-5:] != '.dmnd':
             print('Database must either be a FASTA (.fasta) or a DMND (.dmnd) file')
         else:
+            if not os.path.isfile(database.replace('fasta', 'dmnd')):
+                print('DMND database not found. Generating a new one')
+                self.generate_diamond_database(database,
+                                               database.replace('fasta', 'dmnd'))
             database = database.split('.dmnd')[0]
-        
-        run_command("diamond blastp --query {} --out {} --un {} --db {} --outfmt 6 --unal 1 --threads {} --max-target-seqs {}".format(
-            query, aligned, unaligned, database, threads, max_target_seqs))
-    
+
+        run_command(
+            "diamond blastp --query {} --out {} --un {} --db {} --outfmt 6 --unal 1 --threads {} --max-target-seqs {}".format(
+                query, aligned, unaligned, database, threads, max_target_seqs))
+
     def run(self):
         args = self.get_arguments()
 
-        self.gene_calling(args.input, '{}/fgs'.format(args.output), threads = args.threads,
-                          assembled = args.assembled, error_model = args.error_model)
+        self.gene_calling(args.input, '{}/fgs'.format(args.output), threads=args.threads,
+                          assembled=args.assembled, error_model=args.error_model)
 
-        if args.download_uniprot:
-            self.download_uniprot('/'.join(args.database.split('/')[:-1]))
-        
+        self.download_uniprot('/'.join(args.database.split('/')[:-1]))
+        args.database = '{}/uniprot.fasta'.format('/'.join(args.database.split('/')[:-1]))
+
         self.run_diamond('{}/fgs.faa'.format(args.output), '{}/aligned.blast'.format(args.output),
                          '{}/unaligned.fasta'.format(args.output), args.database,
-                         threads=args.threads, max_target_seqs=args.max_target_seqs)                                  # TODO - annotation has to be refined to retrieve better than hypothetical proteins
+                         threads=args.threads,
+                         max_target_seqs=args.max_target_seqs)  # TODO - annotation has to be refined to retrieve better than hypothetical proteins
 
     '''
     Input:
@@ -120,43 +134,45 @@ class Annotater:
     Output:
         returns data with new protein names
     '''
-    def further_annotation(self, data, temp_folder = 'temp',
-                           dmnd_db = 'MOSCA/Databases/annotation_databases/uniprot.dmnd',
-                           threads = '12', out_format = '6', all_info = True):
+
+    def further_annotation(self, data, temp_folder='temp',
+                           dmnd_db='MOSCA/Databases/annotation_databases/uniprot.dmnd',
+                           threads='12', out_format='6', all_info=True):
         '''
-        noid_entries = data[(data['Protein names']=='Uncharacterized protein') & 
+        noid_entries = data[(data['Protein names']=='Uncharacterized protein') &
                        (data['COG general functional category']=='POORLY CHARACTERIZED')]['Entry']
 
-        self.recursive_uniprot_fasta(temp_folder + '/temp.fasta', 
+        self.recursive_uniprot_fasta(temp_folder + '/temp.fasta',
                                                    entries = noid_entries)
-        
+
         mtools.run_command(('diamond blastp -q {} --db {} -p {} -f {} -o {}').format(
-                temp_folder + '/temp.fasta', dmnd_db, threads, out_format, 
+                temp_folder + '/temp.fasta', dmnd_db, threads, out_format,
                 temp_folder + '/temp_blast.tsv'))
         '''
         blast = mtools.parse_blast(temp_folder + '/temp_blast.tsv')
         '''
         noid_entries = [ide.split('|')[1] for ide in blast['sseqid']]
-        
+
         uniprotinfo = self.get_uniprot_information(noid_entries)
         uniprotinfo.to_csv(temp_folder + '/up.info',index=False)
         '''
         uniprotinfo = pd.read_csv(temp_folder + '/up.info')
-        uniprotinfo.drop_duplicates(inplace = True)
-        
+        uniprotinfo.drop_duplicates(inplace=True)
+
         relation = pd.DataFrame()
         relation['qseqid'] = [ide.split('|')[1] for ide in blast['qseqid']]
         relation['sseqid'] = [ide.split('|')[1] for ide in blast['sseqid']]
-        relation = pd.merge(relation, uniprotinfo, left_on = 'sseqid', 
-                            right_on = 'Entry', how = 'inner')
+        relation = pd.merge(relation, uniprotinfo, left_on='sseqid',
+                            right_on='Entry', how='inner')
         relation = relation[(relation['Protein names'] != 'Deleted.') &
                             (relation['Protein names'] != 'Conserved protein') &
-                            (~relation['Protein names'].str.contains('uncharacterized', case = False))]
-        tax_columns = ['Taxonomic lineage (SUPERKINGDOM)','Taxonomic lineage (PHYLUM)',
-                       'Taxonomic lineage (CLASS)','Taxonomic lineage (ORDER)',
-                       'Taxonomic lineage (FAMILY)','Taxonomic lineage (GENUS)',
+                            (~relation['Protein names'].str.contains('uncharacterized', case=False))]
+        tax_columns = ['Taxonomic lineage (SUPERKINGDOM)', 'Taxonomic lineage (PHYLUM)',
+                       'Taxonomic lineage (CLASS)', 'Taxonomic lineage (ORDER)',
+                       'Taxonomic lineage (FAMILY)', 'Taxonomic lineage (GENUS)',
                        'Taxonomic lineage (SPECIES)']
-        data['Alternative entry'] = np.nan; data['Alternative protein name'] = np.nan
+        data['Alternative entry'] = np.nan;
+        data['Alternative protein name'] = np.nan
         pbar = ProgressBar()
         print('Searching for new annotations in the blast matches.')
         print('Proteins queried for new annotation:', len(set(relation['qseqid'])))
@@ -169,10 +185,12 @@ class Annotater:
             else:
                 if len(partial) > 0:
                     original_taxonomy = data.loc[data['Entry'] == query].iloc[0][tax_columns]
-                    max_proximity = 0; entry = np.nan; protein_name = np.nan
+                    max_proximity = 0;
+                    entry = np.nan;
+                    protein_name = np.nan
                     for i in range(len(partial)):
                         tax_score = self.score_taxonomic_proximity(original_taxonomy,
-                            partial.iloc[i][tax_columns])
+                                                                   partial.iloc[i][tax_columns])
                         if tax_score > max_proximity:
                             max_proximity = tax_score
                             entry = partial.iloc[i]['sseqid']
@@ -189,6 +207,7 @@ class Annotater:
             else:
                 proximity += 1
         return proximity
-    
+
+
 if __name__ == '__main__':
     Annotater().run()
